refactor colors and pages into data
