+++ src/navigation/NavMenu.js
@@ -1,6 +1,6 @@
 import { NavItem } from "components";
+import { colors, pages } from "data";
 import { useActiveNavItem } from "hooks";
-import { pages } from "Main";
 import React from "react";
 import styled from "styled-components";
 
@@ -18,7 +18,7 @@ const ListItem = styled.li`
   transition: linear 0.25s;
 
   padding-bottom: ${(props) => props.active && "5px"};
-  border-bottom: ${(props) => props.active && "2px solid #2bc5e0"};
+  border-bottom: ${(props) => props.active && `2px solid ${colors.blue}`};
 `;
 
 const NavMenu = () => {
